Replace TensorFlow novelty detection with Torch implementation
