+++ src/resources/scripts/novelty-detection/NoveltyDetector.py
